Rename --add-upstream-vcs to --add-upstreamvcs keeping old as alias
The name of the upstream remote when created by git-buildpackage is
'upstreamvcs'. For the sake of consistency, the option that adds it
should be using the same spelling without a dash between 'upstream' and
'vcs'.

Continue to support the old way of spelling the option for backwards
compatibility, but in help and documentation promote the unified
spelling.