Rename --upstream-vcs-tag to --upstreamvcs-tag keeping old as alias
The name of the upstream remote when created by git-buildpackage is
'upstreamvcs'. For the sake of consistency, the option referring to
upstream tags should use the same spelling, without a dash between
'upstream' and 'vcs'.

Continue to support the old way of spelling the option for backward
compatibility, but in help and documentation promote the unified
spelling.